feature: Makes "ConnStandardMaxBufSize" less.
This makes "ConnStandardMaxBufSize" less, because, according to tests,
it doesn't matter if it's 2 times. 65535 works fine on its own.

Anything less than 65535 is dangerous, however. The maximum packet size
in TCP is 64k, so if the maximum buffer size is less than that, it could
cause issues, depending on the application.